requirements.txt
GitPython==3.1.0
unittest2==1.1.0
pytest==4.6.4
pytest-cov==2.5.1
codecov==2.0.17
twine==1.13.0
keyring==19.0.1
PyYAML==5.3.1
jinja2==2.11.1
PyGithub==1.51
truffleHog3==2.0.1
pytest-shutil==1.7.0